fix(6993): further refactor file structure
