+++ development/metamaskbot-build-announce/utils.ts
@@ -5,8 +5,8 @@ import startCase from 'lodash/startCase';
 import {
   BENCHMARK_PLATFORMS,
   BENCHMARK_BUILD_TYPES,
-  STARTUP_PRESETS,
-} from '../../test/e2e/benchmarks/utils/constants';
+} from '../../shared/constants/benchmarks';
+import { STARTUP_PRESETS } from '../../test/e2e/benchmarks/utils/constants';
 import type { BenchmarkResults } from '../../test/e2e/benchmarks/utils/types';
 
 /**
@@ -56,14 +56,12 @@ export type BenchmarkEntry = {
 export type PageLoadEntry = Record<string, Record<string, string>>;
 
 export type PageLoadBenchmarkResults = Record<
-  (typeof BENCHMARK_PLATFORMS)[number],
-  Record<(typeof BENCHMARK_BUILD_TYPES)[number], Record<string, PageLoadEntry>>
+  (typeof BENCHMARK_PLATFORMS)[keyof typeof BENCHMARK_PLATFORMS],
+  Record<
+    (typeof BENCHMARK_BUILD_TYPES)[keyof typeof BENCHMARK_BUILD_TYPES],
+    Record<string, PageLoadEntry>
+  >
 >;
-
-/**
- * The threshold for whether to highlight a change in bundle size, in bytes.
- */
-export const BUNDLE_SIZE_THRESHOLD = 1_000;
 
 /**
  * Converts a byte count to a human-readable string (e.g. "1.5 KiB").
@@ -438,20 +436,14 @@ export async function buildBenchmarkSectionComment(
   const allEntries: BenchmarkEntry[] = [];
 
   for (const preset of presets) {
-    try {
-      const data = await fetchBenchmarkJson(
-        hostUrl,
-        BENCHMARK_PLATFORMS[0],
-        BENCHMARK_BUILD_TYPES[0],
-        preset,
-      );
-      if (data) {
-        allEntries.push(...extractEntries(data));
-      }
-    } catch (error) {
-      console.error(
-        `Failed to fetch benchmark preset "${preset}": ${String(error)}`,
-      );
+    const data = await fetchBenchmarkJson(
+      hostUrl,
+      BENCHMARK_PLATFORMS.CHROME,
+      BENCHMARK_BUILD_TYPES.BROWSERIFY,
+      preset,
+    );
+    if (data) {
+      allEntries.push(...extractEntries(data));
     }
   }
 
@@ -687,4 +679,263 @@ export async function runBenchmarkGate(
   }
 
   return benchmarkGateBody;
+}
+
+/**
+ * Builds the collapsible "Builds ready" artifacts body for the PR comment.
+ *
+ * @param options - Configuration for the artifacts body.
+ * @param options.hostUrl - Base URL for hosted artifacts.
+ * @param options.version - Extension version string (from package.json).
+ * @param options.shortSha - Abbreviated commit hash.
+ * @param options.artifacts - Artifact links from getArtifactLinks.
+ * @param options.postNewBuilds - Whether to include extension build links.
+ * @param options.lavamoatPolicyChanged - Whether to include the LavaMoat viz link.
+ * @returns Collapsible HTML string.
+ */
+export async function buildArtifactsBody({
+  hostUrl,
+  version,
+  shortSha,
+  artifacts,
+  postNewBuilds,
+  lavamoatPolicyChanged,
+}: {
+  hostUrl: string;
+  version: string;
+  shortSha: string;
+  artifacts: ArtifactLinks;
+  postNewBuilds: boolean;
+  lavamoatPolicyChanged: boolean;
+}): Promise<string> {
+  const contentRows: string[] = [];
+
+  if (postNewBuilds) {
+    contentRows.push(...formatBuildLinks(getBuildLinks(hostUrl, version)));
+  }
+
+  if (lavamoatPolicyChanged) {
+    contentRows.push(`lavamoat build viz: ${artifacts.link('depViz')}`);
+  }
+
+  const bundleMarkup = await discoverBundleArtifacts(hostUrl);
+
+  contentRows.push(
+    `bundle size: ${artifacts.link('bundleSizeStats')}`,
+    `interaction-benchmark: ${artifacts.link('interactionStats')}`,
+    `storybook: ${artifacts.link('storybook')}`,
+    `typescript migration: ${artifacts.link('tsMigrationDashboard')}`,
+    artifacts.link('allArtifacts'),
+    `<details>
+       <summary>bundle viz:</summary>
+       ${bundleMarkup}
+     </details>`,
+  );
+
+  const hiddenContent = `<ul>${contentRows
+    .map((row) => `<li>${row}</li>`)
+    .join('\n')}</ul>`;
+
+  return `<details><summary>Builds ready [${shortSha}]</summary>${hiddenContent}</details>\n\n`;
+}
+
+/**
+ * Fetches page load benchmark JSON data for all platform/buildType/preset combos.
+ *
+ * @param hostUrl - Base URL for CI artifacts.
+ * @returns Populated PageLoadBenchmarkResults object.
+ */
+export async function fetchPageLoadResults(
+  hostUrl: string,
+): Promise<PageLoadBenchmarkResults> {
+  const results = {} as PageLoadBenchmarkResults;
+
+  for (const platform of Object.values(BENCHMARK_PLATFORMS)) {
+    results[platform] = {} as PageLoadBenchmarkResults[typeof platform];
+    for (const buildType of Object.values(BENCHMARK_BUILD_TYPES)) {
+      results[platform][buildType] = {};
+      for (const page of Object.values(STARTUP_PRESETS)) {
+        try {
+          const data = await fetchBenchmarkJson<Record<string, PageLoadEntry>>(
+            hostUrl,
+            platform,
+            buildType,
+            page,
+          );
+          // The JSON key is derived from the benchmark file name (e.g. standard-home.ts
+          // → "standardHome"), not from the preset name ("startupStandardHome"). Use
+          // the first (and only) value in the fetched object and store it under the
+          // preset key so the rest of the pipeline stays consistent.
+          const pageResult = data ? Object.values(data)[0] : undefined;
+          if (pageResult) {
+            results[platform][buildType][page] = pageResult;
+          }
+        } catch (error) {
+          console.log(
+            `Failed to fetch page load data for ${platform}/${buildType}/${page}: ${String(error)}`,
+          );
+        }
+      }
+    }
+  }
+
+  return results;
+}
+
+/** The threshold for whether to highlight a change in bundle size, in bytes. */
+const BUNDLE_SIZE_THRESHOLD = 1_000;
+
+/**
+ * Fetches bundle size stats and builds the bundle size diff collapsible section.
+ *
+ * @param artifacts - The artifact links object from getArtifactLinks.
+ * @param mergeBaseCommitHash - The merge base commit hash for comparison.
+ * @returns HTML string for the bundle size diff section, or empty string on error.
+ */
+export async function buildBundleSizeDiffSection(
+  artifacts: ArtifactLinks,
+  mergeBaseCommitHash: string,
+): Promise<string> {
+  const prBundleSizeStatsResponse = await fetch(artifacts.bundleSizeStats.url);
+  if (!prBundleSizeStatsResponse.ok) {
+    throw new Error(
+      `Failed to fetch prBundleSizeStats, status ${prBundleSizeStatsResponse.statusText}`,
+    );
+  }
+  // This annotation narrows the untyped json() result to the known schema of the bundle size stats artifact.
+  const prBundleSizeStats: Record<string, { size: number }> =
+    await prBundleSizeStatsResponse.json();
+
+  const devBundleSizeStatsResponse = await fetch(artifacts.bundleSizeData.url);
+  if (!devBundleSizeStatsResponse.ok) {
+    throw new Error(
+      `Failed to fetch devBundleSizeStats, status ${devBundleSizeStatsResponse.statusText}`,
+    );
+  }
+  // This annotation narrows the untyped json() result to the known schema of the dev bundle size data.
+  const devBundleSizeStats: Record<
+    string,
+    Record<string, number>
+  > = await devBundleSizeStatsResponse.json();
+
+  const bundleParts = ['background', 'ui', 'common'] as const;
+  type BundlePart = (typeof bundleParts)[number];
+
+  const prSizes: Record<BundlePart, number> = {
+    background: prBundleSizeStats.background.size,
+    ui: prBundleSizeStats.ui.size,
+    common: prBundleSizeStats.common.size,
+  };
+
+  const devSizes: Record<BundlePart, number> = {
+    background: 0,
+    ui: 0,
+    common: 0,
+  };
+  for (const part of bundleParts) {
+    devSizes[part] = devBundleSizeStats[mergeBaseCommitHash]?.[part] ?? 0;
+  }
+
+  const diffs: Record<BundlePart, number> = {
+    background: 0,
+    ui: 0,
+    common: 0,
+  };
+  for (const part of bundleParts) {
+    diffs[part] = prSizes[part] - devSizes[part];
+  }
+
+  const sizeDiffRows = bundleParts.map(
+    (part) =>
+      `${part}: ${getHumanReadableSize(diffs[part])} (${getPercentageChange(
+        devSizes[part],
+        prSizes[part],
+      )}%)`,
+  );
+
+  const sizeDiffHiddenContent = `<ul>${sizeDiffRows
+    .map((row) => `<li>${row}</li>`)
+    .join('\n')}</ul>`;
+
+  const sizeDiffBackground = diffs.background + diffs.common;
+  const sizeDiffUi = diffs.ui + diffs.common;
+
+  let sizeDiffWarning: string | undefined;
+  if (
+    sizeDiffBackground > BUNDLE_SIZE_THRESHOLD ||
+    sizeDiffUi > BUNDLE_SIZE_THRESHOLD
+  ) {
+    sizeDiffWarning = `🚨 Warning! Bundle size has increased!`;
+  } else if (
+    sizeDiffBackground < -BUNDLE_SIZE_THRESHOLD ||
+    sizeDiffUi < -BUNDLE_SIZE_THRESHOLD
+  ) {
+    sizeDiffWarning = `🚀 Bundle size reduced!`;
+  }
+
+  const sizeDiffExposedContent = sizeDiffWarning
+    ? `Bundle size diffs [${sizeDiffWarning}]`
+    : `Bundle size diffs`;
+
+  return `<details><summary>${sizeDiffExposedContent}</summary>${sizeDiffHiddenContent}</details>\n\n`;
+}
+
+/**
+ * Uses Github API to post a comment to a PR with `metamaskbot`.
+ *
+ * @param param - Params object.
+ * @param param.commentBody - Comment to post.
+ * @param param.owner - Github repo owner.
+ * @param param.repository - Github repository.
+ * @param param.prNumber - PR Number.
+ * @param [param.optionalLog] - Optional log for extra debug.
+ * @param [param.commentToken] - PR secret comment token.
+ * @returns A promise with response object or null.
+ */
+export async function postCommentWithMetamaskBot({
+  commentBody,
+  owner,
+  repository,
+  prNumber,
+  optionalLog,
+  commentToken,
+}: {
+  commentBody: string;
+  owner: string;
+  repository: string;
+  prNumber: string;
+  optionalLog?: string;
+  commentToken?: string;
+}): Promise<Response | null> {
+  const JSON_PAYLOAD = JSON.stringify({ body: commentBody });
+  const POST_COMMENT_URI = `https://api.github.com/repos/${owner}/${repository}/issues/${prNumber}/comments`;
+
+  if (optionalLog) {
+    console.log(optionalLog);
+  }
+
+  if (!commentToken) {
+    return null;
+  }
+
+  console.log(`Posting to: ${POST_COMMENT_URI}`);
+
+  const response = await fetch(POST_COMMENT_URI, {
+    method: 'POST',
+    body: JSON_PAYLOAD,
+    headers: {
+      'User-Agent': 'metamaskbot',
+      Authorization: `token ${commentToken}`,
+      'Content-Type': 'application/json',
+    },
+  });
+
+  if (!response.ok) {
+    const errorText = await response.text();
+    throw new Error(
+      `Post comment failed with status '${response.statusText}': ${errorText}`,
+    );
+  }
+
+  return response;
 }
